FashionAI/training/requirements_training.txt
# Dependencies used by scripts in FashionAI/FashionAI/training
numpy
opencv-python
matplotlib
pillow
torch
torchvision
